refactor: Simplify the main file and separate into independent modules. (#413)
* refactor: Simplify the main file and separate into independent modules.

* refactor: Optimize file names.

---------
