+++ src/main/setup/auth.ts
@@ -1,0 +1,10 @@
+import { app } from 'electron'
+import { setupTokens } from '../auth'
+
+export default function setupAuth() {
+    app.on('open-url', (_event, url) => {
+        if (url) {
+            setupTokens(url)
+        }
+    })
+}
